refactor: classify internal and infrastructure folder
